requirements.txt
anti-useragent==1.0.10
async-timeout==4.0.2 ; python_full_version <= '3.11.2'
beautifulsoup4==4.11.2
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
colorama==0.4.6 ; sys_platform == 'win32'
deep-translator==1.10.1
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.6.0 ; python_version < '3.8'
jvav==1.2.28
langdetect==1.0.9
loguru==0.7.0 ; python_version >= '3.5'
lxml==4.9.2
pyaes==1.6.1
pyrogram==2.0.92
pysocks==1.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytelegrambotapi==4.10.0
pyyaml==6.0
redis==4.5.5
requests==2.28.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1 ; python_version >= '3.7'
tgcrypto==1.2.5
typing-extensions==4.5.0 ; python_version < '3.8'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wikipedia-api==0.5.8
win32-setctime==1.1.0 ; sys_platform == 'win32'
zipp==3.15.0 ; python_version >= '3.7'
pytest
autopep8
pylint
